bug: Only send notifications when an action has been performed
Previously when an instance went to the stop stage of enforcement, the system would send a notification every `<interval>` minutes, regardless of whether an action had been performed.

Change the system such that it only attempts to stop an instance, if its not already stopped or terminated, and only alert if the action is being taken to remove the notification spam.