Return proper return values :D (#37)
* Return proper return values :D 

The auditor decides whether or not to send a notification based on the return value of the action. In this case we weren't returning False when no action was taken, which lead to the auditor sending constant emails every time it processed a S3 bucket that:
A) already had a lifecycle policy on it
B) already had a bucket policy on it
C) we're waiting to delete pending expiration of the data

This *should* ensure emails are only sent once when the actions are actually taken on the issue.